Refactor circuit breakers and rate limiting

+++ src/tests/test_settings_migration.py
@@ -1,8 +1,8 @@
-﻿import json
+import json
 import os
 from pathlib import Path
 
-from program.settings.manager import SettingsManager
+from program.settings import SettingsManager
 
 TEST_VERSION = "9.9.9"
 DATA_PATH = Path(os.curdir) / "data"
@@ -36,38 +36,29 @@ old_settings_data = {
 }
 
 
-def test_load_and_migrate_settings():
+def test_load_and_migrate_settings(monkeypatch):
     temp_settings_file = Path.joinpath(DATA_PATH, "settings.json")
-    version_file = Path.joinpath(DATA_PATH, "VERSION")
+    DATA_PATH.mkdir(parents=True, exist_ok=True)
 
     try:
         temp_settings_file.write_text(json.dumps(old_settings_data))
-        version_file.write_text("9.9.9")
 
-        import program.settings.models
-
-        program.settings.manager.data_dir_path = DATA_PATH
-        program.settings.models.version_file_path = version_file
+        monkeypatch.setattr("program.settings.data_dir_path", DATA_PATH)
+        monkeypatch.setattr(
+            "program.settings.models.get_version", lambda: TEST_VERSION
+        )
         settings_manager = SettingsManager()
 
-        assert settings_manager.settings.debug is True
-        assert settings_manager.settings.log is True
-        assert settings_manager.settings.force_refresh is False
-        assert settings_manager.settings.map_metadata is True
         assert settings_manager.settings.tracemalloc is False
+        assert settings_manager.settings.logging.enabled is True
         # assert settings_manager.settings.downloaders.movie_filesize_min == 200
         assert settings_manager.settings.downloaders.real_debrid.enabled is False
         assert settings_manager.settings.downloaders.all_debrid.enabled is True
         assert settings_manager.settings.downloaders.all_debrid.api_key == "12345678"
-        assert (
-            settings_manager.settings.downloaders.all_debrid.proxy_url
-            == "https://no_proxy.com"
-        )
-        assert (
-            settings_manager.settings.database.host
-            == "postgresql+psycopg2://postgres:postgres@localhost/riven"
+        assert str(settings_manager.settings.database.host) == (
+            "postgresql+psycopg2://postgres:postgres@localhost/riven"
         )
         assert settings_manager.settings.version == TEST_VERSION
     finally:
-        temp_settings_file.unlink()
-        version_file.unlink()
+        if temp_settings_file.exists():
+            temp_settings_file.unlink()
